+++ src/movement/Pathing.ts
@@ -460,7 +460,7 @@ export class Pathing {
 		}
 		return $.costMatrix(room.name, MatrixTypes.sk, () => {
 			let matrix = this.getDefaultMatrix(room).clone();
-			const avoidRange = 4;
+			const avoidRange = 6;
 			_.forEach(room.keeperLairs, lair => {
 				for (let dx = -avoidRange; dx <= avoidRange; dx++) {
 					for (let dy = -avoidRange; dy <= avoidRange; dy++) {