+++ src/zerg/Swarm.ts
@@ -255,6 +255,10 @@ export class Swarm implements ProtoSwarm { // TODO: incomplete
 			// Creeps travel to their relative formation positions
 			const formationPositions = this.getFormationPositionsFromAnchor(assemblyPoint);
 			for (let creep of this.creeps) {
+				if (creep.hasValidTask) {
+					// Ignore creeps which have tasks (usually getting boosted)
+					continue;
+				}
 				if (allowIdleCombat && creep.room.dangerousPlayerHostiles.length > 0 && !this.hasMaxCreeps) {
 					creep.autoSkirmish(creep.room.name);
 				} else {